fix pointer bug, add test
